feat: ✨ update samrt-terrace, add button-counter-async-fsm

+++ 5-tasks/smart-terrace/src/LightSensorImpl.h
@@ -7,5 +7,5 @@ class LightSensorImpl : public LightSensor {
 
    private:
     int pin;
-    float analogValueToLux(int value);
+    double analogValueToLux(int value);
 };
